// NextMigration/lib/types.ts
export interface User {
    id: string;
    email: string;
    displayName: string;
    img_url: string;
    join_date: string;
    joinNumber: number;
    totalSets: number;
    streak_started: string | null;
    streak_count: number | null;
    streak_last_update: string | null;
    last_login: string;
    publicRole: string;
    verified: boolean;
}

export interface Stats {
    totalsets: number;
    timeLearned: number;
    cardsLearned: number;
}

export interface LastStudied {
    set_id: string;
    last_studied: string;
    title: string;
    Course: string;
    type: string;
    progress: number;
    length: number;
}

export interface UserWithStats extends User {
    stats: UserStats;
    lastTen: LastStudied[];
}

export interface UserListResponse {
    users: User[];
}

export interface HeaderInfo {
    displayName: string;
    email: string;
    streak_count: number | null;
    pfp: string;
}

export interface ClassActivity {
    id: string;
    classroom_id: string;
    user_id: string;
    displayName: string;
    img_url: string;
    set_id: string;
    set_type: string;
    title: string;
    last_seen: string;
}

export interface StartPage {
    lastTen: LastStudied[];
    courses: string[];
    class: ClassActivity[];
}